Request body: $request_body variable generalization.
The $request_body variable was assuming there can't be more than two
buffers.  While this is currently true due to request body reading
implementation details, this is not a good thing to depend on and may
change in the future.

--HG--
extra : convert_revision : svn%3Ac3fe7df1-7212-e011-8a91-001109144009/trunk%404922